requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
diff-match-patch==20230430
Django==5.1.1
django-admin-thumbnails==0.2.9
django-filter==24.3
django-import-export==4.1.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.3
